Expose describo crate builder internals via ref
DescriboCrateBuilder now accepts a ref, which will be set to an objects with
exposed internal objects and methods of the crate builder vue components. The
type of things accessible via the ref are described in
describo-crate-builder.d.ts. Exposing the internals via a ref allows us to
better integrate the crate builder with other parts of the application.